FFmpeg: Add info about regular key frames and scene-cut-based
FROM: https://groups.google.com/a/opencast.org/forum/#!searchin/users/key$20frames$20scene%7Csort:date/users/CzkcVbNGpVE/VRev3sAHBAAJ

Full thread: https://superuser.com/questions/908280/what-is-the-correct-way-to-fix-keyframes-in-ffmpeg-for-dash
Most Helpful answer: https://superuser.com/a/1223359

this is less of a question and more an informational post.
I was adapting our Encoding Profiles to support Adaptive Streaming when I stumbled upon this Superuser Thread. Since it is quite helpful and easy to understand, I want to share it with you.

Full thread: https://superuser.com/questions/908280/what-is-the-correct-way-to-fix-keyframes-in-ffmpeg-for-dash
Most Helpful answer: https://superuser.com/a/1223359

**
** TL;DR: There is a way to get both regular Key Frames for Adaptive Streaming and scene-cut-based Key Frames for better encoding. There are some misconceptions about x264 to be aware of, however.
**

The most important requirement for Adaptive Streaming (AFAIK this is independant of the protocol e.g. DASH/HLS etc) is that the Key Frames of the different quality videos must be at the exact same position since the changes between different qualities can only happen on simultaneous Key Frames.
This is most easily achieved by setting a fixed GOP size of X, so that every X frames there is a Key Frame. When searching the list archive, I found some encoding profiles that are used in production and rely on this method. However, this is not the natural way most encoders work, since they usually take scene cuts into account. These are good spots for Key Frames since a lot of the video content changes. Disabling the scene cut detection and using a fixed rate for Key Frames may result in a scene cut happening in between P or B frames, where mostly motion compensation is used. This is probably not optimal.

This is the topic of the Superuser thread (link at the top), and the last answer (2nd link) explains a possible solution to get both: Regular Key Frames for Adaptive Streaming and scene-cut-based Key Frames for better encoding.
This solution makes use of a 'special kind of key frames', called Non-IDR I-frames. They also contain a complete image and are therefore perfect for frames with scene changes, but they do not start a new GOP, meaning the 'standard' key frames can still occur in regular intervals.

I ran a few tests and compared the results in the table attached to this message. The table shows the frame numbers and types for different encoding settings. You can see that with the encoding settings from the Superuser Thread, the 'standard' key frames are at the exact same frame numbers as with the currently used settings, but there are additional Non-IDR I-Frames at frames where scene changes occur.

I don't know if this is something important when dealing with lecture recordings, since the video material does usually not contain hard cuts and is easy to encode. However I still wanted to leave this here in case someone runs into trouble or needs information on handling key frames. Also, the linked answer in the thread explains that the 'min-keyint' option for x264 does not work as expected and can therefore be left out, as is also shown in the comparison table.

I hope this is of use for someone!