Update PaasSaas.md
Great news, today! Roboflow, the end-to-end annotation and training platform, is now an OpenCV Silver Member. Before this Roboflow was already a member of the community, and you may recognize the name from or their recent OAK tutorial, How to Train and Deploy a License Plate Detector.

Check mail to try de platform for free